requirements.txt
flask==3.0.0
flask-cors==4.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
reportlab>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
jieba>=0.42.1
scikit-learn>=1.3.0
Pillow>=10.0.0
